Changes include both fixes for r9 and general fixes that may want to be back-applied to previous modules.
r9: The shipped versions of compilers have changed. 4.7 for example has been removed. Officially, 4.6 is Android the default, so setting this as the CMake default seems to make more sense.
Added android-18 to the list. I also made the latest the default because there is no blessed default and the most rational default is the latest since that is the whole point of shipping a new NDK release.

General: $ANDROID_NDK_ROOT is the semi-official blessed environmental variable for the location of the root of the Android NDK according to an Android engineer on the official NDK mailing list. For the non-standalone toolchain, this is the directory we want and we don't need a separate $ENV{NDK_TOOLCHAIN_ROOT}. So the logic is now intended to look to the old NDK_TOOLCHAIN_ROOT as an override, and if not set, fallback to the  ANDROID_NDK_ROOT and convenience.

x86: There is a problem changing the arch to x86 after the initial generation. Some values like the compiler and sysroot are stuck on the arm setting. I partly fixed the --sysroot by changing the COMMON_FLAGS to not contain it and put it directly into the flags because the CACHE FORCE flags are complicating things. However, this is not enough to fix the linking flags and I get linker failures (with redundant, conflicting --sysroots)

However, I can successfully generate a project by passing the arch via command line on initial creation.
ccmake -DCMAKE_TOOLCHAIN_FILE=../CMake-Toolchain-Collection/toolchain-android-ndk-r9.cmake -DANDROID_NDK_ARCH=x86 ../MyProject